Added command registry
